+++ backend/canisters/group/impl/src/updates/send_message_v2.rs
@@ -5,16 +5,25 @@ use canister_api_macros::update_candid_and_msgpack;
 use canister_timer_jobs::TimerJobs;
 use canister_tracing_macros::trace;
 use chat_events::{PushMessageArgs, Reader};
-use group_canister::send_message::{Response::*, *};
+use group_canister::send_message_v2::{Response::*, *};
 use std::collections::HashSet;
 use types::{
-    BlobReference, ContentValidationError, EventIndex, EventWrapper, GroupMessageNotification, GroupReplyContext,
-    MentionInternal, Message, MessageContent, MessageIndex, Notification, TimestampMillis, UserId,
+    BlobReference, ContentValidationError, CryptoTransaction, EventIndex, EventWrapper, GroupMessageNotification,
+    GroupReplyContext, MentionInternal, Message, MessageContent, MessageContentInitial, MessageIndex, Notification,
+    TimestampMillis, UserId,
 };
 
 #[update_candid_and_msgpack]
 #[trace]
-fn send_message(args: Args) -> Response {
+fn send_message(args: group_canister::send_message::Args) -> Response {
+    run_regular_jobs();
+
+    mutate_state(|state| send_message_impl(args.into(), state))
+}
+
+#[update_candid_and_msgpack]
+#[trace]
+fn send_message_v2(args: Args) -> Response {
     run_regular_jobs();
 
     mutate_state(|state| send_message_impl(args, state))
@@ -44,7 +53,18 @@ fn send_message_impl(args: Args, runtime_state: &mut RuntimeState) -> Response {
                 ContentValidationError::InvalidTypeForForwarding => {
                     InvalidRequest("Cannot forward this type of message".to_string())
                 }
+                ContentValidationError::PrizeEndDateInThePast => InvalidRequest("Prize ended in the past".to_string()),
             };
+        }
+
+        if let Some(transfer) = match &args.content {
+            MessageContentInitial::Crypto(c) => Some(&c.transfer),
+            MessageContentInitial::Prize(c) => Some(&c.transfer),
+            _ => None,
+        } {
+            if !matches!(transfer, CryptoTransaction::Completed(_)) {
+                return InvalidRequest("The crypto transaction must be completed".to_string());
+            }
         }
 
         let permissions = &runtime_state.data.permissions;
@@ -57,7 +77,7 @@ fn send_message_impl(args: Args, runtime_state: &mut RuntimeState) -> Response {
             return NotAuthorized;
         }
 
-        if matches!(args.content, MessageContent::Poll(_)) && !participant.role.can_create_polls(permissions) {
+        if matches!(args.content, MessageContentInitial::Poll(_)) && !participant.role.can_create_polls(permissions) {
             return NotAuthorized;
         }
 
@@ -220,6 +240,9 @@ fn register_timer_jobs(
             );
         }
     }
+
+    // TODO: If this is a prize message then set a timer to transfer
+    // the balance of any remaining prizes to the original sender
 }
 
 fn get_user_being_replied_to(