+++ backend/canisters/user/api/src/updates/edit_message.rs
@@ -11,11 +11,16 @@ pub struct Args {
     pub correlation_id: u64,
 }
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-    MessageNotFound,
-    ChatNotFound,
-    UserBlocked,
-    UserSuspended,
+pub type Response = crate::edit_message_v2::Response;
+
+impl From<Args> for crate::edit_message_v2::Args {
+    fn from(args: Args) -> Self {
+        crate::edit_message_v2::Args {
+            user_id: args.user_id,
+            thread_root_message_index: args.thread_root_message_index,
+            message_id: args.message_id,
+            content: args.content.into(),
+            correlation_id: args.correlation_id,
+        }
+    }
 }