+++ backend/canisters/user/api/src/updates/send_message_with_transfer_to_group.rs
@@ -1,0 +1,43 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use types::{
+    ChatId, CompletedCryptoTransaction, Cryptocurrency, EventIndex, GroupReplyContext, MessageContentInitial, MessageId,
+    MessageIndex, TimestampMillis, User,
+};
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub group_id: ChatId,
+    pub thread_root_message_index: Option<MessageIndex>,
+    pub message_id: MessageId,
+    pub content: MessageContentInitial,
+    pub sender_name: String,
+    pub replies_to: Option<GroupReplyContext>,
+    pub mentioned: Vec<User>,
+    pub correlation_id: u64,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(SuccessResult),
+    TextTooLong(u32),
+    RecipientBlocked,
+    CallerNotInGroup(Option<CompletedCryptoTransaction>),
+    CryptocurrencyNotSupported(Cryptocurrency),
+    InvalidRequest(String),
+    TransferFailed(String),
+    TransferCannotBeZero,
+    TransferLimitExceeded(u128),
+    UserSuspended,
+    ChatFrozen,
+    InternalError(String, CompletedCryptoTransaction),
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct SuccessResult {
+    pub event_index: EventIndex,
+    pub message_index: MessageIndex,
+    pub timestamp: TimestampMillis,
+    pub expires_at: Option<TimestampMillis>,
+    pub transfer: CompletedCryptoTransaction,
+}