+++ backend/canisters/group/api/src/main.rs
@@ -21,6 +21,7 @@ fn main() {
     generate_candid_method!(group, add_reaction, update);
     generate_candid_method!(group, block_user, update);
     generate_candid_method!(group, change_role, update);
+    generate_candid_method!(group, claim_prize, update);
     generate_candid_method!(group, delete_messages, update);
     generate_candid_method!(group, disable_invite_code, update);
     generate_candid_method!(group, edit_message, update);
@@ -32,6 +33,7 @@ fn main() {
     generate_candid_method!(group, remove_participant, update);
     generate_candid_method!(group, remove_reaction, update);
     generate_candid_method!(group, reset_invite_code, update);
+    generate_candid_method!(group, send_message_v2, update);
     generate_candid_method!(group, send_message, update);
     generate_candid_method!(group, unblock_user, update);
     generate_candid_method!(group, undelete_messages, update);