+++ backend/canisters/notifications/impl/src/lifecycle/post_upgrade.rs
@@ -1,10 +1,11 @@
 use crate::lifecycle::{init_state, UPGRADE_BUFFER_SIZE};
+use crate::memory::get_upgrades_memory;
 use crate::Data;
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk_macros::post_upgrade;
+use ic_stable_structures::reader::{BufferedReader, Reader};
 use notifications_canister::post_upgrade::Args;
-use stable_memory::deserialize_from_stable_memory;
 use tracing::info;
 use utils::cycles::init_cycles_dispenser_client;
 use utils::env::canister::CanisterEnv;
@@ -14,8 +15,10 @@ use utils::env::canister::CanisterEnv;
 fn post_upgrade(args: Args) {
     let env = Box::new(CanisterEnv::new());
 
-    let (data, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>) =
-        deserialize_from_stable_memory(UPGRADE_BUFFER_SIZE).unwrap();
+    let memory = get_upgrades_memory();
+    let reader = BufferedReader::new(UPGRADE_BUFFER_SIZE, Reader::new(&memory, 0));
+
+    let (data, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>) = serializer::deserialize(reader).unwrap();
 
     canister_logger::init_with_logs(data.test_mode, logs, traces);