+++ backend/canisters/user/api/src/updates/send_message.rs
@@ -1,9 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{
-    ChatId, CompletedCryptoTransaction, EventIndex, InvalidPollReason, MessageContent, MessageId, MessageIndex, ReplyContext,
-    TimestampMillis, UserId,
-};
+use types::{MessageContent, MessageId, MessageIndex, ReplyContext, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -17,39 +14,19 @@ pub struct Args {
     pub correlation_id: u64,
 }
 
-#[allow(clippy::large_enum_variant)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success(SuccessResult),
-    TransferSuccessV2(TransferSuccessV2Result),
-    MessageEmpty,
-    TextTooLong(u32),
-    RecipientBlocked,
-    RecipientNotFound,
-    InvalidPoll(InvalidPollReason),
-    InvalidRequest(String),
-    TransferFailed(String),
-    TransferCannotBeZero,
-    TransferLimitExceeded(u128),
-    UserSuspended,
-    InternalError(String),
-}
+pub type Response = crate::send_message_v2::Response;
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub struct SuccessResult {
-    pub chat_id: ChatId,
-    pub event_index: EventIndex,
-    pub message_index: MessageIndex,
-    pub timestamp: TimestampMillis,
-    pub expires_at: Option<TimestampMillis>,
-}
-
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub struct TransferSuccessV2Result {
-    pub chat_id: ChatId,
-    pub event_index: EventIndex,
-    pub message_index: MessageIndex,
-    pub timestamp: TimestampMillis,
-    pub expires_at: Option<TimestampMillis>,
-    pub transfer: CompletedCryptoTransaction,
+impl From<Args> for crate::send_message_v2::Args {
+    fn from(args: Args) -> Self {
+        crate::send_message_v2::Args {
+            recipient: args.recipient,
+            thread_root_message_index: args.thread_root_message_index,
+            message_id: args.message_id,
+            content: args.content.into(),
+            sender_name: args.sender_name,
+            replies_to: args.replies_to,
+            forwarding: args.forwarding,
+            correlation_id: args.correlation_id,
+        }
+    }
 }